+++ zarrs_metadata_ext/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `CastValueCodecConfiguration`, `CastValueCodecConfigurationV1`, `CastValueRoundingMode`, `CastValueOutOfRangeMode`, and `CastValueScalarMap`
+
 ## [0.4.2] - 2026-03-26
 
 ### Added